@@ -41,38 +41,40 @@
     "date-fns": "^2.16.1"
   },
   "devDependencies": {
-    "@angular-devkit/core": "^20.2.0",
-    "@angular-devkit/schematics": "^20.2.0",
-    "@angular/animations": "^20.2.0",
-    "@angular/build": "^20.2.0",
-    "@angular/cli": "^20.2.0",
-    "@angular/common": "^20.2.0",
-    "@angular/compiler": "^20.2.0",
-    "@angular/compiler-cli": "^20.2.0",
-    "@angular/core": "^20.2.0",
-    "@angular/forms": "^20.2.0",
-    "@angular/language-service": "^20.2.0",
-    "@angular/platform-browser": "^20.2.0",
-    "@angular/platform-browser-dynamic": "^20.2.0",
-    "@angular/platform-server": "^20.2.0",
-    "@angular/router": "^20.2.0",
-    "@angular/service-worker": "^20.2.0",
-    "@angular/ssr": "^20.2.0",
+    "@angular-devkit/core": "^20.3.0",
+    "@angular-devkit/schematics": "^20.3.0",
+    "@angular/animations": "^20.3.0",
+    "@angular/build": "^20.3.0",
+    "@angular/cli": "^20.3.0",
+    "@angular/common": "^20.3.0",
+    "@angular/compiler": "^20.3.0",
+    "@angular/compiler-cli": "^20.3.0",
+    "@angular/core": "^20.3.0",
+    "@angular/forms": "^20.3.0",
+    "@angular/language-service": "^20.3.0",
+    "@angular/platform-browser": "^20.3.0",
+    "@angular/platform-browser-dynamic": "^20.3.0",
+    "@angular/platform-server": "^20.3.0",
+    "@angular/router": "^20.3.0",
+    "@angular/service-worker": "^20.3.0",
+    "@angular/ssr": "^20.3.0",
     "@commitlint/cli": "^19.0.0",
     "@commitlint/config-angular": "^19.0.0",
     "@ctrl/tinycolor": "^3.6.0",
     "@eslint/js": "^9.34.0",
     "@eslint/markdown": "^7.0.0",
+    "@inquirer/prompts": "^7.0.0",
     "@schematics/angular": "^20.2.0",
     "@stackblitz/sdk": "^1.11.0",
     "@types/d3": "^6.2.0",
     "@types/express": "^5.0.2",
     "@types/fs-extra": "^11.0.4",
-    "@types/gulp": "^4.0.7",
+    "@types/gulp": "^4.0.17",
     "@types/html-minifier-terser": "^7.0.2",
     "@types/jasmine": "^5.0.0",
     "@types/less": "^3.0.7",
     "@types/lodash": "^4.17.20",
+    "@types/minimist": "^1.2.5",
     "@types/node": "^20.11.1",
     "angular-eslint": "^20.2.0",
     "chalk": "^4.1.0",
@@ -84,6 +86,7 @@
     "d3": "^6.3.1",
     "dagre": "^0.8.5",
     "dagre-compound": "^0.0.8",
+    "del": "^8.0.0",
     "detect-port": "^2.1.0",
     "eslint": "^9.34.0",
     "eslint-plugin-header": "~3.1.1",
@@ -94,7 +97,7 @@
     "eslint-plugin-unused-imports": "^4.1.4",
     "express": "^5.1.0",
     "fs-extra": "^11.3.0",
-    "gulp": "^4.0.2",
+    "gulp": "^5.0.0",
     "gulp-clean": "^0.4.0",
     "html-minifier-terser": "^7.2.0",
     "husky": "^9.0.0",
@@ -118,16 +121,15 @@
     "marked-highlight": "^2.2.2",
     "minimist": "^1.2.5",
     "monaco-editor": "^0.33.0",
-    "ng-packagr": "^20.1.0",
+    "ng-packagr": "^20.3.0",
     "parse5": "^8.0.0",
     "postcss": "^8.5.1",
     "prettier": "^3.6.0",
     "prismjs": "^1.29.0",
     "process": "^0.11.10",
     "puppeteer": "^24.14.0",
-    "readline-sync": "^1.4.10",
     "remark": "^15.0.0",
-    "resolve-bin": "^0.4.0",
+    "resolve-bin": "^1.0.1",
     "rxjs": "~7.8.1",
     "sitemap": "^8.0.0",
     "slugify": "^1.6.6",
