@@ -3,9 +3,10 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-import { ChangeDetectionStrategy, Component, Input, TemplateRef } from '@angular/core';
+import { ChangeDetectionStrategy, Component, input, TemplateRef } from '@angular/core';
 
 import { NzOutletModule } from 'ng-zorro-antd/core/outlet';
+import { NzShapeSCType } from 'ng-zorro-antd/core/types';
 import { NzIconModule } from 'ng-zorro-antd/icon';
 
 @Component({
@@ -16,18 +17,18 @@ import { NzIconModule } from 'ng-zorro-antd/icon';
   template: `
     <div class="ant-float-btn-body">
       <div class="ant-float-btn-content">
-        @if (nzDescription || nzIcon) {
-          @if (nzIcon) {
+        @if (nzDescription() || nzIcon()) {
+          @if (nzIcon()) {
             <div class="ant-float-btn-icon">
-              <ng-container *nzStringTemplateOutlet="nzIcon; let icon">
+              <ng-container *nzStringTemplateOutlet="nzIcon(); let icon">
                 <nz-icon [nzType]="icon" nzTheme="outline" />
               </ng-container>
             </div>
           }
-          @if (nzDescription && nzShape === 'square') {
+          @if (nzDescription() && nzShape() === 'square') {
             <div class="ant-float-btn-description">
-              <ng-container *nzStringTemplateOutlet="nzDescription">
-                {{ nzDescription }}
+              <ng-container *nzStringTemplateOutlet="nzDescription()">
+                {{ nzDescription() }}
               </ng-container>
             </div>
           }
@@ -41,7 +42,7 @@ import { NzIconModule } from 'ng-zorro-antd/icon';
   `
 })
 export class NzFloatButtonContentComponent {
-  @Input() nzIcon: string | TemplateRef<void> | null = null;
-  @Input() nzDescription: string | TemplateRef<void> | null = null;
-  @Input() nzShape: 'circle' | 'square' = 'circle';
+  readonly nzIcon = input<string | TemplateRef<void> | null>(null);
+  readonly nzDescription = input<string | TemplateRef<void> | null>(null);
+  readonly nzShape = input<NzShapeSCType>('circle');
 }
