@@ -8,23 +8,36 @@ import { NzCheckListModule, NzItemProps } from 'ng-zorro-antd/check-list';
   template: `<nz-check-list [nzItems]="nzItems" [nzIndex]="index"></nz-check-list>`
 })
 export class NzDemoCheckListBasicComponent {
-  index = 1;
+  index = 2;
   readonly nzItems: NzItemProps[] = [
     {
       description: 'step 1',
-      onClick: () => this.index++
+      checked: true,
+      onClick: (item: NzItemProps) => {
+        this.index++;
+        item.checked = true;
+      }
     },
     {
       description: 'step 2',
-      onClick: () => this.index++
+      onClick: (item: NzItemProps) => {
+        this.index++;
+        item.checked = true;
+      }
     },
     {
       description: 'step 3',
-      onClick: () => this.index++
+      onClick: (item: NzItemProps) => {
+        this.index++;
+        item.checked = true;
+      }
     },
     {
       description: 'step 4',
-      onClick: () => this.index++
+      onClick: (item: NzItemProps) => {
+        this.index++;
+        item.checked = true;
+      }
     }
   ];
 }
