@@ -3,19 +3,31 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-import { CSP_NONCE, DestroyRef, Injectable, afterNextRender, assertInInjectionContext, inject } from '@angular/core';
+import {
+  CSP_NONCE,
+  DestroyRef,
+  Injectable,
+  afterNextRender,
+  assertInInjectionContext,
+  inject,
+  Signal,
+  computed,
+  WritableSignal,
+  signal,
+  InputSignalWithTransform,
+  InputSignal
+} from '@angular/core';
+import { SIGNAL } from '@angular/core/primitives/signals';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 import { Observable, Subject, Subscription } from 'rxjs';
 import { filter, map } from 'rxjs/operators';
 
-import { NzSafeAny } from 'ng-zorro-antd/core/types';
-
 import { NZ_CONFIG, NzConfig, NzConfigKey } from './config';
 import { registerTheme } from './css-variables';
 
-const isDefined = function (value?: NzSafeAny): boolean {
+function isDefined<T>(value?: T): value is Exclude<T, undefined> {
   return value !== undefined;
-};
+}
 
 const defaultPrefixCls = 'ant';
 
@@ -24,6 +36,13 @@ const defaultPrefixCls = 'ant';
 })
 export class NzConfigService {
   private configUpdated$ = new Subject<keyof NzConfig>();
+
+  /**
+   * Sharing config signals for all components, used for {@link withConfigFactory}
+   * @internal
+   * @todo use nested signal to refactor the whole config service
+   */
+  private readonly _configMap = new Map<NzConfigKey, WritableSignal<NzConfig[NzConfigKey]>>();
 
   /** Global config holding property. */
   private readonly config: NzConfig = inject(NZ_CONFIG, { optional: true }) || {};
@@ -35,6 +54,17 @@ export class NzConfigService {
       // If theme is set with NZ_CONFIG, register theme to make sure css variables work
       registerTheme(this.getConfig().prefixCls?.prefixCls || defaultPrefixCls, this.config.theme, this.cspNonce);
     }
+  }
+
+  private _getConfigValue<T extends NzConfigKey>(componentName: T): WritableSignal<NzConfig[T]> {
+    let configValue = this._configMap.get(componentName) as WritableSignal<NzConfig[T]>;
+    if (configValue) {
+      return configValue;
+    }
+
+    configValue = signal(this.config[componentName]);
+    this._configMap.set(componentName, configValue);
+    return configValue;
   }
 
   getConfig(): NzConfig {
@@ -54,6 +84,7 @@ export class NzConfigService {
 
   set<T extends NzConfigKey>(componentName: T, value: NzConfig[T]): void {
     this.config[componentName] = { ...this.config[componentName], ...value };
+    this._configMap.get(componentName)?.set(this.config[componentName]);
     if (componentName === 'theme' && this.config.theme) {
       registerTheme(this.getConfig().prefixCls?.prefixCls || defaultPrefixCls, this.config.theme, this.cspNonce);
     }
@@ -141,6 +172,53 @@ export function WithConfig<This, Value>() {
         enumerable: true,
         configurable: true
       });
+    });
+  };
+}
+
+/**
+ * Generate a `withConfig` function for a specific component, which would try to load default value from `NZ_CONFIG`
+ * if the `input` property is not assigned by user.
+ *
+ * @param componentName The name of component (as defined in {@link NzConfigKey}) to listen for config changes.
+ * @example
+ * ```ts
+ * const withConfig = withConfigFactory('button');
+ *
+ * class ExampleComponent {
+ *   readonly nzSize = input<NzButtonSize>('default');
+ *   protected readonly size = withConfig('nzSize', this.nzSize);
+ * }
+ * ```
+ */
+export function withConfigFactory<T extends NzConfigKey>(componentName: T) {
+  /**
+   * @param name The name of input property.
+   * @param inputSignal The input signal.
+   * @param defaultValue The default value.
+   */
+  return <N extends keyof NonNullable<NzConfig[T]>, V = NonNullable<NzConfig[T]>[N]>(
+    name: N,
+    inputSignal: InputSignal<V | undefined> | InputSignalWithTransform<V | undefined, unknown>,
+    defaultValue: V
+  ): Signal<V> => {
+    const configValueSignal = inject(NzConfigService)['_getConfigValue'](componentName);
+
+    return computed<V>(() => {
+      const configValue = configValueSignal()?.[name] as V | undefined;
+      const inputValue = inputSignal();
+      // if the version of the inputSignal is 0 or the inputValue is undefined, we consider it as not assigned by user
+      const assignedByUser = inputSignal[SIGNAL].version > 0 && isDefined(inputValue);
+
+      if (assignedByUser) {
+        return inputValue;
+      }
+
+      if (isDefined(configValue)) {
+        return configValue;
+      }
+
+      return defaultValue;
     });
   };
 }
