@@ -3,24 +3,26 @@
  * found in the LICENSE file at https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/LICENSE
  */
 
-import { Direction, Directionality } from '@angular/cdk/bidi';
+import { Directionality } from '@angular/cdk/bidi';
 import {
   ChangeDetectionStrategy,
-  ChangeDetectorRef,
   Component,
-  DestroyRef,
-  EventEmitter,
+  computed,
   inject,
-  Input,
-  OnInit,
-  Output,
+  input,
+  linkedSignal,
+  output,
   TemplateRef
 } from '@angular/core';
-import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 
 import { NzButtonModule } from 'ng-zorro-antd/button';
+import { NzShapeSCType } from 'ng-zorro-antd/core/types';
+import { generateClassName } from 'ng-zorro-antd/core/util';
 
 import { NzFloatButtonContentComponent } from './float-button-content.component';
+import { NzFloatButtonType } from './typings';
+
+const CLASS_NAME = 'ant-float-btn';
 
 @Component({
   selector: 'nz-float-button',
@@ -28,65 +30,63 @@ import { NzFloatButtonContentComponent } from './float-button-content.component'
   imports: [NzButtonModule, NzFloatButtonContentComponent],
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
-    @if (!!nzHref) {
+    @if (!!nzHref()) {
       <a
-        [target]="nzTarget"
-        [href]="nzHref"
+        [target]="nzTarget()"
+        [href]="nzHref()"
         nz-button
-        [nzType]="nzType"
-        [class.ant-float-btn-default]="nzType === 'default'"
+        [nzType]="nzType()"
+        [class.ant-float-btn-default]="nzType() === 'default'"
         class="ant-float-btn-inner"
         (click)="nzOnClick.emit(true)"
       >
         <nz-float-button-content
-          [nzIcon]="nzIcon"
-          [nzDescription]="nzDescription"
-          [nzShape]="nzShape"
+          [nzIcon]="nzIcon()"
+          [nzDescription]="nzDescription()"
+          [nzShape]="shape()"
         ></nz-float-button-content>
       </a>
     } @else {
       <button
         nz-button
-        [nzType]="nzType"
-        [class.ant-float-btn-default]="nzType === 'default'"
+        [nzType]="nzType()"
+        [class.ant-float-btn-default]="nzType() === 'default'"
         class="ant-float-btn-inner"
         (click)="nzOnClick.emit(true)"
       >
         <nz-float-button-content
-          [nzIcon]="nzIcon"
-          [nzDescription]="nzDescription"
-          [nzShape]="nzShape"
+          [nzIcon]="nzIcon()"
+          [nzDescription]="nzDescription()"
+          [nzShape]="shape()"
         ></nz-float-button-content>
       </button>
     }
   `,
   host: {
-    class: 'ant-float-btn',
-    '[class.ant-float-btn-circle]': `nzShape === 'circle'`,
-    '[class.ant-float-btn-square]': `nzShape === 'square'`,
-    '[class.ant-float-btn-rtl]': `dir === 'rtl'`
+    '[class]': 'class()'
   }
 })
-export class NzFloatButtonComponent implements OnInit {
-  private directionality = inject(Directionality);
-  private cdr = inject(ChangeDetectorRef);
-  private destroyRef = inject(DestroyRef);
+export class NzFloatButtonComponent {
+  readonly nzHref = input<string | null>(null);
+  readonly nzTarget = input<string | null>(null);
+  readonly nzType = input<NzFloatButtonType>('default');
+  readonly nzIcon = input<string | TemplateRef<void> | null>(null);
+  readonly nzDescription = input<string | TemplateRef<void> | null>(null);
+  readonly nzShape = input<NzShapeSCType>('circle');
+  readonly nzOnClick = output<boolean>();
 
-  @Input() nzHref: string | null = null;
-  @Input() nzTarget: string | null = null;
-  @Input() nzType: 'default' | 'primary' = 'default';
-  @Input() nzShape: 'circle' | 'square' = 'circle';
-  @Input() nzIcon: string | TemplateRef<void> | null = null;
-  @Input() nzDescription: TemplateRef<void> | string | null = null;
-  @Output() readonly nzOnClick = new EventEmitter<boolean>();
-  dir: Direction = 'ltr';
+  readonly shape = linkedSignal(() => this.nzShape());
+  protected readonly dir = inject(Directionality).valueSignal;
+  protected readonly class = computed<string[]>(() => {
+    const dir = this.dir();
+    const classes = [CLASS_NAME, this.generateClass(this.shape())];
+    if (dir === 'rtl') {
+      classes.push(this.generateClass(dir));
+    }
+    return classes;
+  });
 
-  ngOnInit(): void {
-    this.directionality.change?.pipe(takeUntilDestroyed(this.destroyRef)).subscribe(direction => {
-      this.dir = direction;
-      this.cdr.detectChanges();
-    });
-
-    this.dir = this.directionality.value;
+  private generateClass(suffix: string): string {
+    return generateClassName(CLASS_NAME, suffix);
   }
 }
