@@ -4,15 +4,12 @@
  */
 
 import { Component, inject, OnDestroy } from '@angular/core';
-import { ComponentFixture, TestBed, inject as testingInject } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { Subscription } from 'rxjs';
 
 import { NZ_I18N, provideNzI18n } from 'ng-zorro-antd/i18n/nz-i18n.token';
 
-import cs_CZ from './languages/cs_CZ';
-import de_DE from './languages/de_DE';
 import en_US from './languages/en_US';
-import ka_GE from './languages/ka_GE';
 import zh_CN from './languages/zh_CN';
 import { NzI18nInterface } from './nz-i18n.interface';
 import { NzI18nService } from './nz-i18n.service';
@@ -23,30 +20,15 @@ describe('i18n service', () => {
   let testComponent: NzI18nTestComponent;
   const DEFAULT_LAN = zh_CN;
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [provideNzI18n(DEFAULT_LAN)]
-    });
-  });
-
   describe('#setLocale', () => {
     beforeEach(() => {
+      TestBed.configureTestingModule({
+        providers: [provideNzI18n(DEFAULT_LAN)]
+      });
+
       fixture = TestBed.createComponent(NzI18nTestComponent);
       testComponent = fixture.debugElement.componentInstance;
-    });
-
-    beforeEach(
-      testingInject([NzI18nService], (s: NzI18nService) => {
-        srv = s;
-      })
-    );
-
-    it('should interface be right', () => {
-      const i18nEN: NzI18nInterface = en_US;
-      const i18nDE: NzI18nInterface = de_DE;
-      const i18nCS: NzI18nInterface = cs_CZ;
-      const i18nKA: NzI18nInterface = ka_GE;
-      console.log(i18nEN, i18nDE, i18nCS, i18nKA);
+      srv = TestBed.inject(NzI18nService);
     });
 
     it('should be provide interface be right', () => {
@@ -83,6 +65,22 @@ You can use "NzI18nService.setLocale" as a temporary fix.
 Welcome to submit a pull request to help us optimize the translations!
 https://github.com/NG-ZORRO/ng-zorro-antd/blob/master/CONTRIBUTING.md`
       );
+    });
+  });
+
+  describe('provideNzI18n', () => {
+    beforeEach(() => {
+      TestBed.configureTestingModule({
+        providers: [provideNzI18n(() => en_US)]
+      });
+
+      fixture = TestBed.createComponent(NzI18nTestComponent);
+      testComponent = fixture.debugElement.componentInstance;
+      srv = TestBed.inject(NzI18nService);
+    });
+
+    it('should factory work', () => {
+      expect(testComponent.locale.locale).toBe(en_US.locale);
     });
   });
 });
