@@ -7,9 +7,10 @@ import { Direction, Directionality } from '@angular/cdk/bidi';
 import { DOWN_ARROW, ENTER, ESCAPE, LEFT_ARROW, RIGHT_ARROW, TAB, UP_ARROW } from '@angular/cdk/keycodes';
 import {
   ConnectionPositionPair,
+  createFlexibleConnectedPositionStrategy,
+  createOverlayRef,
+  createRepositionScrollStrategy,
   FlexibleConnectedPositionStrategy,
-  Overlay,
-  OverlayConfig,
   OverlayRef,
   PositionStrategy
 } from '@angular/cdk/overlay';
@@ -21,6 +22,7 @@ import {
   ChangeDetectionStrategy,
   ChangeDetectorRef,
   Component,
+  computed,
   contentChild,
   ContentChild,
   DestroyRef,
@@ -29,6 +31,7 @@ import {
   ElementRef,
   EventEmitter,
   inject,
+  Injector,
   Input,
   NgZone,
   OnChanges,
@@ -36,6 +39,7 @@ import {
   Output,
   QueryList,
   Renderer2,
+  signal,
   SimpleChanges,
   TemplateRef,
   ViewChild,
@@ -47,8 +51,9 @@ import { merge, of as observableOf, Subscription } from 'rxjs';
 import { distinctUntilChanged, map, startWith, switchMap, withLatestFrom } from 'rxjs/operators';
 
 import { NzFormItemFeedbackIconComponent, NzFormNoStatusService, NzFormStatusService } from 'ng-zorro-antd/core/form';
+import { NzStringTemplateOutletDirective } from 'ng-zorro-antd/core/outlet';
 import { DEFAULT_MENTION_BOTTOM_POSITIONS, DEFAULT_MENTION_TOP_POSITIONS } from 'ng-zorro-antd/core/overlay';
-import { NgClassInterface, NzSafeAny, NzStatus, NzValidateStatus } from 'ng-zorro-antd/core/types';
+import { NgClassInterface, NzSafeAny, NzStatus, NzValidateStatus, NzVariant } from 'ng-zorro-antd/core/types';
 import {
   fromEventOutsideAngular,
   getCaretCoordinates,
@@ -118,19 +123,39 @@ export type MentionPlacement = 'top' | 'bottom';
     @if (hasFeedback && !!status) {
       <nz-form-item-feedback-icon class="ant-mentions-suffix" [status]="status" />
     }
+    @if (nzAllowClear && hasValue()) {
+      <span class="ant-mentions-suffix">
+        <button type="button" tabindex="-1" class="ant-mentions-clear-icon" (click)="clear()">
+          <ng-template [nzStringTemplateOutlet]="nzClearIcon">
+            <nz-icon nzType="close-circle" nzTheme="fill" />
+          </ng-template>
+        </button>
+      </span>
+    }
   `,
   changeDetection: ChangeDetectionStrategy.OnPush,
   host: {
     class: 'ant-mentions',
-    '[class.ant-mentions-rtl]': `dir === 'rtl'`
+    '[class.ant-mentions-rtl]': `dir === 'rtl'`,
+    '[class.ant-mentions-borderless]': `nzVariant === 'borderless'`,
+    '[class.ant-mentions-filled]': `nzVariant === 'filled'`,
+    '[class.ant-mentions-underlined]': `nzVariant === 'underlined'`,
+    '[class.ant-mentions-focused]': `focused()`,
+    '[class.ant-mentions-disabled]': `disabled()`
   },
-  imports: [NgTemplateOutlet, NzIconModule, NzEmptyModule, NzFormItemFeedbackIconComponent]
+  imports: [
+    NgTemplateOutlet,
+    NzIconModule,
+    NzEmptyModule,
+    NzFormItemFeedbackIconComponent,
+    NzStringTemplateOutletDirective
+  ]
 })
 export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
   private ngZone = inject(NgZone);
   private directionality = inject(Directionality);
   private cdr = inject(ChangeDetectorRef);
-  private overlay = inject(Overlay);
+  private injector = inject(Injector);
   private viewContainerRef = inject(ViewContainerRef);
   private elementRef = inject(ElementRef);
   private renderer = inject(Renderer2);
@@ -142,8 +167,12 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
   @Input() nzPlacement: MentionPlacement = 'bottom';
   @Input() nzSuggestions: NzSafeAny[] = [];
   @Input() nzStatus: NzStatus = '';
+  @Input() nzVariant: NzVariant = 'outlined';
+  @Input({ transform: booleanAttribute }) nzAllowClear = false;
+  @Input() nzClearIcon: TemplateRef<NzSafeAny> | null = null;
   @Output() readonly nzOnSelect = new EventEmitter<NzSafeAny>();
   @Output() readonly nzOnSearchChange = new EventEmitter<MentionOnSearchTypes>();
+  @Output() readonly nzOnClear = new EventEmitter<void>();
 
   @ViewChild(TemplateRef, { static: false }) suggestionsTemp?: TemplateRef<void>;
   @ViewChildren('items', { read: ElementRef })
@@ -168,6 +197,14 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
   statusCls: NgClassInterface = {};
   status: NzValidateStatus = '';
   hasFeedback: boolean = false;
+  readonly focused = signal(false);
+
+  readonly disabled = computed(() => {
+    return this.trigger().disabled();
+  });
+  readonly hasValue = computed(() => {
+    return !!this.trigger()?.value().trim();
+  });
 
   private previousValue: string | null = null;
   private cursorMention: string | null = null;
@@ -272,7 +309,7 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
   }
 
   getMentions(): string[] {
-    return this.trigger() ? getMentions(this.trigger().value!, this.nzPrefix) : [];
+    return this.trigger() ? getMentions(this.trigger().value(), this.nzPrefix) : [];
   }
 
   selectSuggestion(suggestion: string | {}): void {
@@ -287,10 +324,16 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
     this.activeIndex = -1;
   }
 
+  clear(): void {
+    this.closeDropdown();
+    this.trigger().clear();
+    this.nzOnClear.emit();
+  }
+
   private handleInput(event: KeyboardEvent): void {
     const target = event.target as HTMLInputElement | HTMLTextAreaElement;
     this.trigger().onChange(target.value);
-    this.trigger().value = target.value;
+    this.trigger().value.set(target.value);
     this.resetDropdown();
   }
 
@@ -327,6 +370,8 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
   }
 
   private bindTriggerEvents(): void {
+    this.trigger().onFocusin.subscribe(() => this.focused.set(true));
+    this.trigger().onBlur.subscribe(() => this.focused.set(false));
     this.trigger().onInput.subscribe((e: KeyboardEvent) => this.handleInput(e));
     this.trigger().onKeydown.subscribe((e: KeyboardEvent) => this.handleKeydown(e));
     this.trigger().onClick.subscribe(() => this.handleClick());
@@ -469,7 +514,11 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
   private attachOverlay(): void {
     if (!this.overlayRef) {
       this.portal = new TemplatePortal(this.suggestionsTemp!, this.viewContainerRef);
-      this.overlayRef = this.overlay.create(this.getOverlayConfig());
+      this.overlayRef = createOverlayRef(this.injector, {
+        positionStrategy: this.getOverlayPosition(),
+        scrollStrategy: createRepositionScrollStrategy(this.injector),
+        disposeOnNavigation: true
+      });
     }
     if (this.overlayRef && !this.overlayRef.hasAttached()) {
       this.overlayRef.attach(this.portal);
@@ -478,26 +527,14 @@ export class NzMentionComponent implements OnInit, AfterViewInit, OnChanges {
     this.updatePositions();
   }
 
-  private getOverlayConfig(): OverlayConfig {
-    return new OverlayConfig({
-      positionStrategy: this.getOverlayPosition(),
-      scrollStrategy: this.overlay.scrollStrategies.reposition(),
-      disposeOnNavigation: true
-    });
-  }
-
   private getOverlayPosition(): PositionStrategy {
-    const positions = [
-      new ConnectionPositionPair({ originX: 'start', originY: 'bottom' }, { overlayX: 'start', overlayY: 'top' }),
-      new ConnectionPositionPair({ originX: 'start', originY: 'top' }, { overlayX: 'start', overlayY: 'bottom' })
-    ];
-    this.positionStrategy = this.overlay
-      .position()
-      .flexibleConnectedTo(this.trigger().elementRef)
-      .withPositions(positions)
+    return (this.positionStrategy = createFlexibleConnectedPositionStrategy(this.injector, this.trigger().elementRef)
+      .withPositions([
+        new ConnectionPositionPair({ originX: 'start', originY: 'bottom' }, { overlayX: 'start', overlayY: 'top' }),
+        new ConnectionPositionPair({ originX: 'start', originY: 'top' }, { overlayX: 'start', overlayY: 'bottom' })
+      ])
       .withFlexibleDimensions(false)
-      .withPush(false);
-    return this.positionStrategy;
+      .withPush(false));
   }
 
   private setStatusStyles(status: NzValidateStatus, hasFeedback: boolean): void {
