@@ -7,6 +7,7 @@ import { AnimationEvent } from '@angular/animations';
 import { Directionality } from '@angular/cdk/bidi';
 import { DOWN_ARROW, LEFT_ARROW, RIGHT_ARROW, UP_ARROW } from '@angular/cdk/keycodes';
 import {
+  afterNextRender,
   booleanAttribute,
   ChangeDetectionStrategy,
   ChangeDetectorRef,
@@ -17,6 +18,7 @@ import {
   EventEmitter,
   forwardRef,
   inject,
+  Injector,
   Input,
   OnChanges,
   Output,
@@ -73,6 +75,7 @@ const NZ_CONFIG_MODULE_NAME: NzConfigKey = 'segmented';
     '[class.ant-segmented-sm]': `nzSize === 'small'`,
     '[class.ant-segmented-block]': `nzBlock`,
     '[class.ant-segmented-vertical]': `nzVertical`,
+    '[class.ant-segmented-shape-round]': `nzShape === 'round'`,
     // a11y
     role: 'radiogroup',
     'aria-label': 'segmented control',
@@ -94,13 +97,15 @@ export class NzSegmentedComponent implements OnChanges, ControlValueAccessor {
 
   public readonly nzConfigService = inject(NzConfigService);
   private readonly cdr = inject(ChangeDetectorRef);
-  protected readonly dir = inject(Directionality).valueSignal;
   private readonly service = inject(NzSegmentedService);
+  private readonly injector = inject(Injector);
+  protected readonly dir = inject(Directionality).valueSignal;
 
   @Input({ transform: booleanAttribute }) nzBlock: boolean = false;
   @Input({ transform: booleanAttribute }) nzDisabled = false;
   @Input() nzOptions: NzSegmentedOptions = [];
   @Input({ transform: booleanAttribute }) nzVertical: boolean = false;
+  @Input() nzShape: 'default' | 'round' = 'default';
   @Input() @WithConfig() nzSize: NzSizeLDSType = 'default';
 
   // todo: add a method to generate hash id for the segmented instance as default value of `nzName`
@@ -170,19 +175,24 @@ export class NzSegmentedComponent implements OnChanges, ControlValueAccessor {
       )
       .subscribe(event => this.onKeyDown(event));
 
-    effect(() => {
-      const itemCmps = this.renderedItemCmps();
+    afterNextRender(() => {
+      effect(
+        () => {
+          const itemCmps = this.renderedItemCmps();
 
-      if (!itemCmps.length) {
-        return;
-      }
+          if (!itemCmps.length) {
+            return;
+          }
 
-      if (
-        this.value === undefined || // If no value is set, select the first item
-        !itemCmps.some(item => item.nzValue() === this.value) // handle value not in options
-      ) {
-        this.service.selected$.next(itemCmps[0].nzValue());
-      }
+          if (
+            this.value === undefined || // If no value is set, select the first item
+            !itemCmps.some(item => item.nzValue() === this.value) // handle value not in options
+          ) {
+            this.service.selected$.next(itemCmps[0].nzValue());
+          }
+        },
+        { injector: this.injector }
+      );
     });
   }
 
