@@ -5,4 +5,4 @@
 
 import { Version } from '@angular/core';
 
-export const VERSION = new Version('20.2.1');
+export const VERSION = new Version('20.3.1');
