@@ -8,9 +8,10 @@ import { getProjectFromWorkspace } from '@angular/cdk/schematics';
 import { chain, noop, Rule, schematic, SchematicContext, Tree } from '@angular-devkit/schematics';
 import { NodePackageInstallTask, RunSchematicTask } from '@angular-devkit/schematics/tasks';
 import { readWorkspace } from '@schematics/angular/utility';
+import { addPackageJsonDependency, NodeDependencyType } from '@schematics/angular/utility/dependencies';
+import { latestVersions } from '@schematics/angular/utility/latest-versions';
 
 import { Schema } from './schema';
-import { addPackageToPackageJson } from '../utils/package-config';
 import { getProjectStyle } from '../utils/project-style';
 // generated by scripts/schematics/set-version.ts
 // @ts-ignore
@@ -25,9 +26,16 @@ export default function(options: Schema): Rule {
       // command, or Material is only listed a dev dependency. If that is the case, we insert a
       // version based on the current build version (substituted version placeholder).
       if (!options.skipPackageJson) {
-        addPackageToPackageJson(host, 'ng-zorro-antd', zorroVersion);
+        addPackageJsonDependency(host, { name: 'ng-zorro-antd', version: zorroVersion, type: NodeDependencyType.Default });
         if (options.gestures) {
-          addPackageToPackageJson(host, 'hammerjs', hammerjsVersion);
+          addPackageJsonDependency(host, { name: 'hammerjs', version: hammerjsVersion, type: NodeDependencyType.Default });
+        }
+
+        // Since Angular CLI v20 no longer installs `less` for the project which prefers `css` or 
+        // `scss` as default stylesheet format, and ng-zorro-antd's custom theming requires `less`,
+        // we should add `less` as a dev dependency.
+        if (options.theme) {
+          addPackageJsonDependency(host, { name: 'less', version: latestVersions['less'], type: NodeDependencyType.Dev });
         }
       }
 
